pyproject.toml
[project]
name = "udata-fastapi-test"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.43",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest>=8.4.1",
]
